Adds functional test and facade for shownoter
There is now a functional test to test the
generation of markdown from a simulated chat history

Also added a function that will act as a single entry point
that the internals of shownoter can resude behind.

This will allow easier api manipulatuion as the current api
will be hidden behind the new facade.